+++ src/main/java/org/realityforge/guiceyloops/server/MockPersistenceTestModule.java
@@ -6,22 +6,16 @@ import org.realityforge.guiceyloops.shared.AbstractModule;
 public class MockPersistenceTestModule
   extends AbstractModule
 {
-  private final boolean _bindWithoutName;
   private final String _persistenceUnit;
 
-  public MockPersistenceTestModule( final String persistenceUnit, final boolean bindWithoutName )
+  public MockPersistenceTestModule( final String persistenceUnit )
   {
-    _bindWithoutName = bindWithoutName;
     _persistenceUnit = persistenceUnit;
   }
 
   @Override
   protected void configure()
   {
-    if ( _bindWithoutName )
-    {
-      bindMock( EntityManager.class );
-    }
     if ( null != _persistenceUnit )
     {
       bindMock( EntityManager.class, _persistenceUnit );